Closes #32. Fixed LSL#0. Fixed BIC. Fixed Data Operations with R15 as destination register. Fixed Halfword Data Transfers. Fixed Branch. Fixed Thumb Decoder switch. Fixed Thumb Push/Pop registers. Fixed Thumb Long Branch & Link. Added stricter memory enforcement for invalid BIOS reads.

+++ src/arm7tdmi/types.hpp
@@ -10,6 +10,6 @@ using BitSegment = std::pair<std::uint8_t, std::uint8_t>;
 using ParamSegments = std::vector<BitSegment>;
 
 // Params should be stored from LSB to MSB
-using ParamList = std::vector<std::uint16_t>;
+using ParamList = std::vector<std::uint32_t>;
 
 }  // namespace ARM7TDMI
